requirements.txt
pytest==8.3.3
selenium==4.24.0
pytest-html==4.1.1
PyYAML==6.0.2